requirements.txt
dnspython
flask
flask_cors
joblib
netifaces
pandas
psutil
scapy
scikit-learn